Cleanup network settings when using systemd.
When using systemd-networkd to configure interfaces, they are not
de-configured as the initramfs exits. This means they are in use, and
the udev renaming of the network device names will not work. This adds a
systemd unit and accompanying script that will make sure to bring down
all network interfaces as the initramfs is exiting.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>